requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
gunicorn==20.1.0
django-environ>=0.8.1
django-filter>=2.4.0
Markdown>=3.1
graphene-django==3.1.2
django-cors-headers>=3.5.0
requests==2.31.0
graphene-django-filter==0.6.5
pytest==6.2.4
beautifulsoup4==4.12.3
playwright==1.44.0